Fully implemented 4 pulse magic

+++ Assets/Scripts/Damage/Config/ContinuousEffectiveDamagerConfig.cs
@@ -3,5 +3,5 @@
 [CreateAssetMenu(fileName = "Damager Data", menuName = "ScriptableObjects/ContinuousEffectiveDamagerConfig", order = 1)]
 public class ContinuousEffectiveDamagerConfig : DamagerConfig
 {
-    public int DamageInterval;
+    public float DamageInterval;
 }
